[codespaces] Add some clauses

+++ init-codespace.sh
@@ -1,10 +1,16 @@
 #!/usr/bin/env bash
 
-cp .env.example .env
+if [[ ! -f ".env" ]]; then
+    cp .env.example .env
+fi
 
 composer install
 
-php artisan key:generate
+grep -q "base64:.*=" ./.env; i=$(($?));
+if [[ $i != "0" ]]; then
+    php artisan key:generate
+fi
+
 php artisan migrate:fresh --seed
 
 npm install
